requirements.txt
astroid==2.9.3
click==8.0.3
colorama==0.4.4
Flask==2.0.2
Flask-HTTPAuth==4.7.0
iso8601==1.0.2
isort==5.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
lazy-object-proxy==1.7.1
MarkupSafe==2.0.1
mccabe==0.6.1
mockito==1.3.0
mongomock==4.0.0
packaging==21.3
platformdirs==2.5.0
pylint==2.12.2
pymongo==3.12.1
pyparsing==3.0.7
python-dateutil==2.8.2
sentinels==1.0.0
six==1.16.0
toml==0.10.2
typing-extensions==4.2.0
Werkzeug==2.0.2
wrapt==1.13.3
